feat(client): add Socket context to app
